Adding feature to not let log file grow
Added feature to address feature request #12.

On every startup, the add-on will look for an output.json in the logs directly. Which is enabled by default with my rtl_433.conf

If output.json exists, it will look to see if the file is greater than 1MB. If it is then it will forcefully move it to output.json.bak.

On startup if both exist, it will overwrite the existing output.json bak.